Update bot.py
From lines 23-26 added a quit command that allows us to quit the bot so that we can change code and adjust to bugs before running it again. this fixes the problem of the bot returning the same phrase multiple times after a command. So before fixing code make sure to use the quit command to stop all instances.